ai/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
Werkzeug==2.3.1
